1. Fix vector in AlgoCPU
2. Rename context_manager file

3. Move example functions to main.py
