+++ memory_reference/include/sync_open.py
